+++ docs/conf.py
@@ -26,7 +26,7 @@ project = u'pymailup'
 year = '2016'
 author = u'Thunder Systems'
 copyright = '{0}, {1}'.format(year, author)
-version = release = u'0.3.0'
+version = release = u'0.3.5'
 
 pygments_style = 'trac'
 templates_path = ['.']