+++ src/mailup/__init__.py
@@ -7,4 +7,4 @@ import exceptions
 import providers
 import utils
 
-__version__ = "0.3.0"
+__version__ = "0.3.5"